Fix build by extracting types + const
from useMimirHalt hook

- [AppView] Revert removing `rdAltOnPending` in `renderHaltedChainsWarning`
