+++ src/renderer/components/deposit/add/SymDeposit.stories.tsx
@@ -22,9 +22,9 @@ import * as RxOp from 'rxjs/operators'
 import { mockValidatePassword$ } from '../../../../shared/mock/wallet'
 import { ZERO_BASE_AMOUNT } from '../../../const'
 import { BNB_DECIMAL } from '../../../helpers/assetHelper'
-import { DEFAULT_MIMIR_HALT } from '../../../hooks/useMimirHalt'
 import { INITIAL_SYM_DEPOSIT_STATE } from '../../../services/chain/const'
 import { SymDepositState } from '../../../services/chain/types'
+import { DEFAULT_MIMIR_HALT } from '../../../services/thorchain/const'
 import { WalletBalance } from '../../../types/wallet'
 import { SymDeposit, Props as SymDepositProps } from './SymDeposit'
