+++ src/renderer/components/deposit/add/SymDeposit.tsx
@@ -36,7 +36,6 @@ import { sequenceSOption, sequenceTOption } from '../../../helpers/fpHelpers'
 import * as PoolHelpers from '../../../helpers/poolHelper'
 import { liveData, LiveData } from '../../../helpers/rx/liveData'
 import { FundsCap } from '../../../hooks/useFundsCap'
-import { MimirHalt } from '../../../hooks/useMimirHalt'
 import { useSubscriptionState } from '../../../hooks/useSubscriptionState'
 import { INITIAL_SYM_DEPOSIT_STATE } from '../../../services/chain/const'
 import { getZeroSymDepositFees } from '../../../services/chain/fees'
@@ -54,7 +53,7 @@ import {
 import { OpenExplorerTxUrl } from '../../../services/clients'
 import { ApproveFeeHandler, ApproveParams, IsApprovedRD, LoadApproveFeeHandler } from '../../../services/ethereum/types'
 import { PoolAddress, PoolsDataMap } from '../../../services/midgard/types'
-import { PendingAssets, PendingAssetsRD } from '../../../services/thorchain/types'
+import { MimirHalt, PendingAssets, PendingAssetsRD } from '../../../services/thorchain/types'
 import { ApiError, TxHashLD, TxHashRD, ValidatePasswordHandler } from '../../../services/wallet/types'
 import { AssetWithDecimal } from '../../../types/asgardex'
 import { WalletBalances } from '../../../types/wallet'
