+++ src/renderer/components/deposit/withdraw/Withdraw.stories.tsx
@@ -17,9 +17,9 @@ import * as RxOp from 'rxjs/operators'
 
 import { mockValidatePassword$ } from '../../../../shared/mock/wallet'
 import { BNB_DECIMAL, THORCHAIN_DECIMAL } from '../../../helpers/assetHelper'
-import { DEFAULT_MIMIR_HALT } from '../../../hooks/useMimirHalt'
 import { INITIAL_WITHDRAW_STATE } from '../../../services/chain/const'
 import { WithdrawState$ } from '../../../services/chain/types'
+import { DEFAULT_MIMIR_HALT } from '../../../services/thorchain/const'
 import { Withdraw, Props as WitdrawProps } from './Withdraw'
 
 const defaultProps: WitdrawProps = {
