+++ src/renderer/components/deposit/withdraw/Withdraw.tsx
@@ -25,7 +25,6 @@ import { getTwoSigfigAssetAmount, THORCHAIN_DECIMAL, to1e8BaseAmount } from '../
 import { eqAsset } from '../../../helpers/fp/eq'
 import * as PoolHelpers from '../../../helpers/poolHelper'
 import { liveData } from '../../../helpers/rx/liveData'
-import { MimirHalt } from '../../../hooks/useMimirHalt'
 import { useSubscriptionState } from '../../../hooks/useSubscriptionState'
 import { INITIAL_WITHDRAW_STATE } from '../../../services/chain/const'
 import { getZeroWithdrawFees } from '../../../services/chain/fees'
@@ -39,6 +38,7 @@ import {
 } from '../../../services/chain/types'
 import { OpenExplorerTxUrl } from '../../../services/clients'
 import { PoolsDataMap } from '../../../services/midgard/types'
+import { MimirHalt } from '../../../services/thorchain/types'
 import { ValidatePasswordHandler } from '../../../services/wallet/types'
 import { AssetWithDecimal } from '../../../types/asgardex'
 import { PasswordModal } from '../../modal/password'
