+++ src/renderer/components/swap/Swap.stories.tsx
@@ -21,9 +21,9 @@ import * as RxOp from 'rxjs/operators'
 import { mockValidatePassword$ } from '../../../shared/mock/wallet'
 import { ONE_BN } from '../../const'
 import { THORCHAIN_DECIMAL } from '../../helpers/assetHelper'
-import { DEFAULT_MIMIR_HALT } from '../../hooks/useMimirHalt'
 import { INITIAL_SWAP_STATE } from '../../services/chain/const'
 import { SwapState } from '../../services/chain/types'
+import { DEFAULT_MIMIR_HALT } from '../../services/thorchain/const'
 import { AssetWithDecimal } from '../../types/asgardex'
 import { Swap, SwapProps } from './Swap'
