+++ src/renderer/components/wallet/assets/AssetsTableCollapsable.tsx
@@ -16,10 +16,10 @@ import { Network } from '../../../../shared/api/types'
 import { disableRuneUpgrade, isNonNativeRuneAsset } from '../../../helpers/assetHelper'
 import { getChainAsset } from '../../../helpers/chainHelper'
 import { getPoolPriceValue } from '../../../helpers/poolHelper'
-import { MimirHaltRD } from '../../../hooks/useMimirHalt'
 import * as walletRoutes from '../../../routes/wallet'
 import { WalletBalancesRD } from '../../../services/clients'
 import { PoolDetails } from '../../../services/midgard/types'
+import { MimirHaltRD } from '../../../services/thorchain/types'
 import { ApiError, ChainBalance, ChainBalances } from '../../../services/wallet/types'
 import { WalletBalance, WalletBalances } from '../../../types/wallet'
 import { PricePool } from '../../../views/pools/Pools.types'
