+++ src/renderer/views/pools/Pools.types.ts
@@ -3,7 +3,7 @@ import { BaseAmount, Asset, Chain } from '@xchainjs/xchain-util'
 import { NonEmptyArray } from 'fp-ts/lib/NonEmptyArray'
 
 import { Network } from '../../../shared/api/types'
-import { MimirHalt } from '../../hooks/useMimirHalt'
+import { MimirHalt } from '../../services/thorchain/types'
 import { GetPoolsStatusEnum } from '../../types/generated/midgard'
 
 export type Pool = {
